Module membrane::governance

Modules

helpers

Structs

Config

This structure stores general parameters for the Governance contract.

InstantiateMsg

This structure holds the parameters used for creating a Governance contract.

Proposal

This structure stores data for a proposal.

ProposalListResponse

This structure describes a proposal list response.

ProposalMessage

This structure describes a proposal message.

ProposalResponse

This structure describes a proposal response.

ProposalVote

This structure describes a proposal vote.

ProposalVotesResponse

This structure describes a proposal vote response.

UpdateConfig

This structure stores the params used when updating the main Governance contract params.

Enums

ExecuteMsg

This enum describes all execute functions available in the contract.

ProposalStatus

This enum describes available statuses/states for a Proposal.

ProposalVoteOption

This enum describes available options for voting on a proposal.

QueryMsg

Thie enum describes all the queries available in the contract.

Constants

DELAY_INTERVAL

EXPIRATION_PERIOD_INTERVAL

MAX_PROPOSAL_REQUIRED_QUORUM_PERCENTAGE

MAX_PROPOSAL_REQUIRED_THRESHOLD_PERCENTAGE

MINIMUM_PROPOSAL_REQUIRED_QUORUM_PERCENTAGE

MINIMUM_PROPOSAL_REQUIRED_THRESHOLD_PERCENTAGE

STAKE_INTERVAL

VOTING_PERIOD_INTERVAL