Module membrane::helpers

Constants

SECONDS_PER_YEAR

Functions

accumulate_interest

Accumulate interest to a given base amount

assert_sent_native_token_balance

Asserts balance of native tokens sent to the contract Refactored Terraswap function.

asset_to_coin

Converts Asset to Coin

get_asset_liquidity

Returns asset liquidity from the liquidity check contract

get_contract_balances

Get contract balances for list of assets

get_owner_liquidity_multiplier

Return liquidity multiplier & SP cap ratio for an owner of the Osmosis Proxy contract

get_pool_state_response

Returns PoolStateResponse from Osmosis proxy

multi_native_withdrawal_msg

Builds withdraw msg for multiple native assets

native_asset_to_coin

Converts native Asset to Coin

pool_query_and_exit

Query Osmosis proxy for pool state then create & return LP withdraw msg

query_stability_pool_fee

Returns Stability Pool liq premium

router_native_to_native

Creates router swap msg between native assets

validate_position_owner

Returns valid addr for contract usage

withdrawal_msg

Build withdraw msg for native assets